BTA-2006 Add sender_id as a query parameter to the currencies endpoint (#120)
The BFX app uses the rates information  from the currency info API with the option to get custom rates based on the sender id.
This commit adds this to the specification documents.

It also adds the add_document endpoint to transaction requests

This is needed in the private swagger specs and the private SDK so we can use it for Project Square