BTA-2650 Add recipient proof of payment endpoint (#160)
Add the recipient proof of payment endpoint to the private endpoint
definitions to ensure it can be consumed in Square via the Ruby SDK.

This commit also reverts a previous change to the OpenApi definitions as they were
generated with a version of the openapi-generator different to the one
used in the deploy script here. The version used to generate the files
here is v4.0.0-beta3.

Resolves: https://btcafrica.atlassian.net/browse/BTA-2650