BTA-2780 Update documentation for sender documents (#170)
Rather than use the documents controller to add documents to a sender as well as
delete documents previously added by a sender, we are now adding the method to the
sender model the same way we do in the transaction request model.
This allows us to manage the polymorphic document model without having to handle
scoping of the parent model withing the documents controller.

With this change we have removed the delete endpoint from documents and removed the
sender_id parameter as well as this was misleading. Adding a sender_id parameter raised
the question around also adding the transaction request id parameter.