BTA-2709 add transaction estimate to sender fields (#171)
The number_monthly_transactions and amount_monthly_transactions are part of the data
collected during the onboarding process and this updates the sender endpoints with
these fields.

The state property was added very recently but is only visible with a transaction request.
The functionality is not required by our internal apps right now so is being removed.