BTA-5884 Remove WARI as a documented payment provider (#251)
Due to this payment rail failing for several weeks we are looking to
disable it and migrate our customers to our other XOF Cash provider. Due
to keeping this a backward compatable change we need to retain wari as a
valid enum option which is why it has not been removed in this pull
request.