BTA-5456 Update version numbers, JS Travis template (#255)
Updates the working version numbers to 1.14.0 now that version 1.13.0
has been release.

The Javascript SDK builds were failing due to issues with the wrong
version of NPM being used for builds using older versions of Node. This
update ensures that the latest version of NPM is being used for each
version of Node.

Related to: transferzero/transferzero-sdk-javascript#9